tasks/medplus/freelance_management_system/freelance/requirements.txt
amqp==2.5.2
angular-django==0.0.1
anyio==3.6.1
apturl==0.5.2
arabic-reshaper==2.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.3.0
astroid==2.4.2
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
billiard==3.6.3.0
bleach==5.0.0
blinker==1.4
boto==2.49.0
Brlapi==0.7.0
bs4==0.0.1
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
click==8.1.3
colorama==0.4.3
command-not-found==0.3
cors==1.0.1
cryptography==2.7
cssselect2==0.6.0
cupshelpers==1.0
cycler==0.11.0
dataclasses==0.6
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
device-detector==0.9
dill==0.3.5.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==1.0.0
Django==2.2.4
django-auth-ldap==2.0.0
django-cors-headers==3.1.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-debug-toolbar==2.2
django-filter==22.1
django-heroku==0.3.1
django-ipware==2.1.0
django-on-heroku==1.1.2
django-pdfkit==0.3.1
django-phone-field==1.8.1
django-phonenumber-field==7.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djongo==1.2.31
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.3
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.7.1
fonttools==4.33.3
future==0.18.2
futures==3.0.5
gevent==21.12.0
goslate==1.5.4
greenlet==1.1.3
html5lib==1.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==5.8.0
inflect==4.1.0
iniconfig==1.1.1
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.7.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyter-server==1.17.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.1
keyring==18.0.1
kiwisolver==1.4.3
kombu==4.6.8
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
mysqlclient==1.4.4
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.0
oauthlib==3.1.0
olefile==0.46
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
passlib==1.7.2
pdfkit==0.6.0
pexpect==4.6.0
phonenumbers==8.12.55
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.6.1
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6
pycups==1.9.73
PyDictionary==2.0.1
pydub==0.23.1
pygame==2.1.2
Pygments==2.12.0
PyGObject==3.36.0
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
pymongo==3.9.0
PyNaCl==1.3.0
pyotp==2.4.1
pyparsing==3.0.9
PyPDF2==1.26.0
PyPDF3==1.0.6
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.2
python-apt==2.0.0+ubuntu0.20.4.8
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-ldap==3.2.0
python-pam==1.8.4
python-xapp==1.8.1
python-xlib==0.23
pytimeparse==1.1.8
pytube==12.1.0
pytube3==9.6.4
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==23.1.0
qrcode==7.3.1
qtconsole==5.3.1
QtPy==2.1.0
redis==3.4.1
regex==2020.2.20
reportlab==3.6.11
requests==2.22.0
requests-file==1.5.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.8.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
stack-data==0.3.0
svglib==1.4.1
systemd-python==234
terminado==0.15.0
tinycss==0.4
tinycss2==1.1.1
tk==0.1.0
tldextract==3.3.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
typed-ast==1.4.2
typing_extensions==4.2.0
tzdata==2022.2
tzlocal==4.2
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritools==4.0.0
urllib3==1.25.8
usb-creator==0.3.7
vine==1.3.0
virtualenv==20.15.1
virtualenvwrapper-win==1.2.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
whitenoise==6.2.0
widgetsnbextension==3.6.1
wkhtmltopdf==0.2
wrapt==1.12.1
xhtml2pdf==0.2.8
xkit==0.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zipstream-new==1.1.8
zope.event==4.5.0
zope.interface==5.4.0
